wrap auth callback in suspense
