+++ src/app/auth/callback/CallbackClient.tsx
@@ -1,0 +1,43 @@
+'use client'
+
+import { useEffect } from 'react'
+import { useRouter, useSearchParams } from 'next/navigation'
+import { supabase } from '@/lib/supabaseClient'
+
+const ADMIN_EMAILS = ['rezvalia@gmail.com', 'jozef.bubliak@gmail.com']
+
+export default function CallbackClient() {
+  const router = useRouter()
+  const params = useSearchParams()
+
+  useEffect(() => {
+    const handleRedirect = async () => {
+      // krátka pauza, kým sa stihne vytvoriť session po návrate z OAuth
+      await new Promise((resolve) => setTimeout(resolve, 1000))
+
+      const {
+        data: { session },
+      } = await supabase.auth.getSession()
+      const user = session?.user
+
+      console.log('[CALLBACK] Session:', session)
+      console.log('[CALLBACK] User:', user)
+
+      const next = params.get('next')
+      if (next) {
+        router.push(next)
+        return
+      }
+
+      if (user?.email && ADMIN_EMAILS.includes(user.email)) {
+        router.push('/admin')
+      } else {
+        router.push('/app')
+      }
+    }
+
+    handleRedirect()
+  }, [params, router])
+
+  return <p className="text-center p-10 text-gray-500">Prihlasovanie...</p>
+}
