+++ src/app/auth/callback/page.tsx
@@ -1,43 +1,12 @@
-'use client'
+import { Suspense } from 'react'
+import CallbackClient from './CallbackClient'
 
-import { useEffect } from 'react'
-import { useRouter, useSearchParams } from 'next/navigation'
-import { supabase } from '@/lib/supabaseClient'
+export const dynamic = 'force-dynamic'
 
-const ADMIN_EMAILS = ['rezvalia@gmail.com', 'jozef.bubliak@gmail.com']
-
-export default function AuthCallback() {
-  const router = useRouter()
-  const params = useSearchParams()
-
-  useEffect(() => {
-    const handleRedirect = async () => {
-      // krátka pauza, kým sa stihne vytvoriť session po návrate z OAuth
-      await new Promise((resolve) => setTimeout(resolve, 1000))
-
-      const {
-        data: { session },
-      } = await supabase.auth.getSession()
-      const user = session?.user
-
-      console.log('[CALLBACK] Session:', session)
-      console.log('[CALLBACK] User:', user)
-
-      const next = params.get('next')
-      if (next) {
-        router.push(next)
-        return
-      }
-
-      if (user?.email && ADMIN_EMAILS.includes(user.email)) {
-        router.push('/admin')
-      } else {
-        router.push('/app')
-      }
-    }
-
-    handleRedirect()
-  }, [params, router])
-
-  return <p className="text-center p-10 text-gray-500">Prihlasovanie...</p>
+export default function AuthCallbackPage() {
+  return (
+    <Suspense fallback={null}>
+      <CallbackClient />
+    </Suspense>
+  )
 }
